Trim shop background state parameter

+++ src/Event/Fcl/fclShopBGManager.c
@@ -41,7 +41,7 @@ u32 FUN_00410e70(u32 param_1,int param_2,int param_3);
 u32 FUN_004110f0(u64 param_1,u64 param_2,int param_3);
 u32 FUN_00411190(u64 param_1,u64 param_2,u32 param_3);
 u32 FUN_00411340(u64 param_1,u64 param_2,s16 *param_3);
-u32 FUN_004113f0(u64 param_1,int param_2,long param_3);
+u32 FUN_004113f0(u64 param_1,int param_2,int param_3);
 u64 FUN_00411710(u64 param_1,int param_2);
 u32 FUN_00411790(u64 param_1,int param_2,long param_3);
 u64 FUN_00411ee0(u64 param_1,int param_2);
@@ -1050,7 +1050,7 @@ u32 FUN_00411340(u64 param_1,u64 param_2,s16 *param_3)
 // FUN_004113F0 NONMATCHING
 
 
-u32 FUN_004113f0(u64 param_1,int param_2,long param_3)
+u32 FUN_004113f0(u64 param_1,int param_2,int param_3)
